Add initial implementation
This is the result of the initial development/MVP
for Next REST Framework. The framework already
includes extensive features like support for OpenAPI
spec auto generations, object-schema support for Zod
and Yup and lots of user config options but the API
is not considered to be fully stable yet.

This includes the monorepo contents with the actual
library code, a development application, CI configs
etc. that are needed for further development.