v0.1.0
This is the result of the initial development/MVP
for Next REST Framework and includes the release
v0.1.0. The framework already includes extensive features
like support for OpenAPIspec auto generations,
object-schema support for Zod and Yup and lots of
user config options but the API is not considered
to be fully stable yet.

This includes the monorepo contents with the actual
library code, a development application, CI configs
etc. that are needed for further development.