v0.1.1
This includes the following improvements
and release v0.1.1:

- Fix NPM package not working due to CommonJS and
ESM interoperability issue.
- Fix missing `summary` field from OpenAPI paths.
- Fix miscellaneous typos in README.

- Add an example application with the installed NPM package.

Additionally, the core logic for instrumenting the API routes
for the OpenAPI definitions is changed quite a bit by introducing
additional logic for providing a custom response in the `defineEndpoints`
function that is being used internally for the OpenAPI spec generation
now. This is a much cleaner way to get the configuration of the individual
API routes.

Codecov is also enabled now in the CI pipeline for increased coverage
monitoring.