v0.2.0
This change breaks the API by changing the previous logic
used for both input/output validation and the OpenAPI generation
so that only a simple default specification is generated from the newly
introfuced `input` and `output` logic, whereas the user can override
these with the `openApiSpec` object that can now be used on three
levels, similarly to the middlewares and error handlers.

This change also contains miscellaneous refactorings for the project
structure due to the increasing amount of utility functions etc.