Fix errors with thrid-party API routes
This fixes some compability issues with user-defined
API routes that are not using Next REST Framework.
There was a bug in the API route path matching where
the comparison to the reserved paths was not exact,
making some user-defined API routes to be completely
skipped in some cases.

Other fix here is that the instrumentation requests are now
aborted after 200ms in case the user-defined APIs do not
respond in that time.

The development application now uses the `src` folder
structure with an additional random API route that helped
discovering this issue.