Drop support for Node 14
The recent addition of abort controller usage
caused an additional bug for using the framework
with Node version 14.x.x. Fixing this would require
us to add additional polyfill, most likely a new dependency.
From the maintainer's perspective, that is not feasable in this
state of the project and adding support for Node 14 can be
considered later if there will be demand for that.

This being said, the existing polyfills in the project are also
patching features in the Node 14 runtime, meaning that
after this change they are no longer needed.