Fix error with middlewares that respond to API requests
This fixes a bug with middlewares that responded to the API
request. Previously the endpoint execution was continued after
every middleware and error handler, resulting in an error in the
previously mentioned case.

This fixes the issue by checking wheter the request has been
ended after proceeding after every middleware and error handler.