Add support for local OpenAPI specification
This changes how the framework can be used in production.
Previously, we dynamically re-created the OpenAPI spec every
time one of the OpenAPI endpoints were called.

This chagnes that logic by caching the generated OpenAPI spec
to a local JSON file that is dynamically updated during local development
but never in production.

This change also allows to run the framework in production in serverless
environments as the generated OpenAPI spec is always static and ready
to go.

This also fixes the example application so we can use it as a demo application
and test the framework live in production on every change.