Drop support for Yup schemas
The community have recently pointed out multiple bugs with
Zod schemas and based on feedback the need Yup support is
marginal/non-existent. With the amount of bugs pointed out with Zod schemas,
those same bugs also lie in the Yup schema support. In order to make sure the
experience with Zod schemas is smooth and bug-free, maintaining the support
for both Yup and Zod simultaneously is becoming harder over time.

This change removes the support for Yup schemas completely in order
to focus on fixing the issues with generating the OpenAPI spec from
Zod schemas and improving the testing coverage on those.