Drop yup support & improve Zod schema generation (#47)
* Drop support for Yup schemas

The community have recently pointed out multiple bugs with
Zod schemas and based on feedback the need Yup support is
marginal/non-existent. With the amount of bugs pointed out with Zod schemas,
those same bugs also lie in the Yup schema support. In order to make sure the
experience with Zod schemas is smooth and bug-free, maintaining the support
for both Yup and Zod simultaneously is becoming harder over time.

This change removes the support for Yup schemas completely in order
to focus on fixing the issues with generating the OpenAPI spec from
Zod schemas and improving the testing coverage on those.

* Improve Zod schema generation

This modifies the Zod schema conversion logic by introducing
new supported Zod schema types for the OpenAPI generation
from the Zod documentation: https://zod.dev

Now all schema types that are applicable for a REST API use
case should be supported. The different schema types are also
included in the tests so that all supported schema types should
have coverage with different variations.

The Zod dependecy is also bumped to its latest version.