Add support for App Router
This is a breaking change that adds support for
using Next REST Framework with the Next.js App Router.

The support for Pages Router is also supported, and even
both routers can be used simultaneously with the framework,
like done in the revamped example application.

There are a few fundamental changes in both the implementation
of the code and in how the framework is encouraged to be used:

Remove support for middlewares:

Next.js has it's own way of implementing middleware logic
for server-side code. By removing the optional logic for
global, route and method middlewares we ensure that any
user-defined logic in these handler's won't interfere or
encourage developers to implement their preferred logic in
a non-Next.js way.

Supporting different runtimes, like the Node.js environment and
the edge runtime are also not possible, making it wiser to leave
any additional middleware logic up to the user to implement either
directly in their route handlers, or in their global Next.js
middleware logic.

Leaving our the three-level middleware logic also simplifies the
typings, that are already very strict.

With these changes, the user can still define an optional global
error handler for catching any errors occurring in their Next REST
Framework route handlers, but defining overriding error handlers on
route or method level is no longer supported.

Remove support for HTTP TRACE:

The Next.js App Router adds it's own validation for the HTTP methods
being used, and TRACE is not supported by that. With these changes,
TRACE can no longer be used with neither App Router or Pages Router
with Next REST Framework.

Not encouraging running the generation script with `next build`:

The `next-rest-framework` binary still works in a similar way,
generating the `openapi.json` file in the root of the project.
Running the script that spawns a child process running a Next.js
dev server is longer encouraged to be done with `next build`,
as performing this in build environments can have varying success
or it may not work at all. Instead, it is recommended to still keep
the `openapi.json` file under source control and instead, run the
generation script as part of pre-commit hooks or before committing
changes that modify the OpenAPI spec. The `openapi.json` file
is still automatically updated during local development, making
this manual step redundant in most cases.

Removing support for Node.js 16:

As Node.js 16 is facing EOL, and the Next.js App Router is not
supported with Node.js 16, that Node version is longer supported.
Node 20 can be considered to be added to the CI machine matrix, but
as Vercel does not currently support Node 20, that has not yet been
done here.

Additionally, this change upgrades most dependencies to latest versions.