Fix directory not found error when using App Router
This fixes an issue where an error occurred on the pages
directory not being found when using App Router - this both
removes the default config option that sets a default value
for the `apiRoutesPath` option, and also gracefully handles
the error when either `appDirPath` or `apiRoutesPath` option
is enabled but either of those directories are not found.

Also, this fixes typos in the docs where the App Router catch-all
handler was not exported as an HTTP method.