Fix request body validation
When using App Route and the NextRequest
to validate the request body, we need to clone
the request object to validate the body to allow
further parsing of the request body. Before this
change, trying to parse the request body in the
request handler resulted in an error if the request
body was already parsed by the request body
validation logic.

Also, this updates the example todos endpoints
to match with the examples provided in the readme
and docs.