Fix `appDirPath` handling when not using src directory
There is a bug when not using the src directory and using
the App Router - if the user does not define the `appDirPath`
config option having a slash as prefix, the framework won't
be able to find the route handlers. This fixes that with a one-line
change in the route discovery logic.

The typings for both `appDirPath` and `apiRoutesPath` options
are also changed so that the following formats are now supported:
- `src/app`
- `/src/app`
- `./src/app`