Fix query parameter validation on App Router (#73)
This fixes the search query parameter validation
when using App Router. The previous implementation
that was ported from the Pages Router implementation,
was validating the context parameter, used as part of
the dynamic route parameters with App Router handlers.
This fix makes sure that the query parameter validation
is equivalent on both routers now. The typings for the
request object on App Router are also updated now so that
the typed query parameters can be accessed from
`req.nextUrl.search`.

Co-authored-by: Mirko Budszuhn <mirko314@users.noreply.github.com>