Fix content type validation (#74)
This fix changes the content type validation
so that it's done by the framework only when
the user has defined their desired content type
in the inputs. This should address some unexpected
errors with various API clients that automatically set
the content type header when the user has not set any
header validations themselves.