Add support for strongly-typed NextResponse (#91)
Previously, the problem with App Router was that
truly strongly-typed responses were not possible due
to the `NextResponse` class not supporting a generic
typing of the status codes. This change adds a patched
`TypedNextResponse` class that adds support for
type-checking the response status codes based on
the output schema.

In order to achieve this, the user must import the
`TypedNextResponse` class from Next REST Framework,
but the good thing is that this is fully optional and using the
plain `NextResponse` is still supported if the user prefers not
to have strongly typed status code checks.