Implement RPC API handlers and client
This introduces a new way of writing RPC styled API
handlers and abstract away the use of status codes, HTTP
responses etc. The input validation, type-cheking and OpenAPI
spec generation are handled in the same way as with the REST
endpoint handlers.

Additionally, a new RPC client function is introduced that allows
calling the procedures both on the server and in the client.