Add Edge runtime compatibility (#119)
This is a major change that simplifies the OpenAPI sepc
generation process, and adds Edge runtime support.

Removing development-server generation:

This change completely removes the automatic
local OpenAPI spec generation when calling the
documentation endpoint and leaves the spec
generation completely for the CLI which is now
much more reliable for the following reason:

Previously the file system based OpenAPI
spec generation done by the CLI used the
build output generated by Next.js in the
`.next` folder. This was however suboptimal
as there's no guarantee that the module
structure won't change between Next.js versions,
breaking the CLI. A new approach is that the
CLI commands `generate` and `validate` now
generate a temporary build folder called
`.next-rest-framework` which is compiled using
ESBuild which produces consistent output from
the CLI. This temporary folder is used to gather
the OpenAPI paths from the generated routes
and api routes that works with all Next.js versions.
This is also much faster than running `next build`
every time with the CLI.

With this changes, all Node.js file system based
API calls etc. are split do a different bundle so that
the code also works in the Edge runtime.

Fixed:
- Fixed the ESM entry point causing NRF not
being loaded as ESM.
- Fixed the possible return status codes from RPC
routes to match the generated OpenAPI spec.