Fix CLI not working with explicit module type
This fixes a bug where the CLI commands failed
to compile the endpoints when the client application's
package.json contains an explicit "type": "module".

While the output is still bundled to CJS, the extensions
now make sure the CLI Node.js process can call the
generated modules and generate the OpenAPI paths.

The esbuild plugin for TS alias paths is also not needed
as the CLI Node.js process does not have to resolve any
imports in the compiled endpoints.