Fix query parameter  OpenAPI generation
The OpenAPI spec generation for query
parameters ended up adding unnecessary
union types for the generated schema which
is fixed here.

The query parameter parsing is now done with
the `qs` library instead of the previous manual
parsing and some test coverage is added to handle
both valid and invalid query parameter calls.

Co-authored-by: itsjxck <1731116+itsjxck@users.noreply.github.com>