Fix CLI bundling glob for npm/npx
The ESBuild process for the CLI commands
fails when run with npm/npx due to the dependencies
being resolved differently than with pnpm, resulting in
the dependencies at node_modules folder being resolved
by ESBuild when running the CLI commands with npm/npx.

To fix this, the `tiny-glob` package is replaced with an alternative
`fast-glob` that allows us to explicitly exclude all node_modules
files from the globbing output, fixing the issue.