Add routes with external dependency to example app
This allows testing that API routes with third-party dependencies
like JSDom don't cause issues in the OpenAPI generation.

This commit also adds an example on writing custom scripts
for generating/validating the OpenAPI spec.

The deep object comparison as part of the OpenAPI
generation/validation is also replaced with a faster
stringified comparison between the generated and
existing OpenAPI specs.

Co-authored-by: Austin Kelleher <austinlkelleher@gmail.com>