Fix dynamic server usage in app router docs route (#168)
Currently in the app router docs route we read the request
headers directly from the original request object. This is
problematic because it leads to a dynamic server usage
error unless the docs endpoint is explicitly set to `force-dynamic`
caching strategy. By reading the headers from a cloned request
object the docs endpoint can be used with the default caching
strategy, fixing this issue.