Silence `findConfig` errors
A custom error handling/logging was previously
added to log any errros for importing routes
during the OpenAPI path generation. For finding
the correct docs handler to pull the NRF config
from, that handling is not needed and it may often
result to a lot of visible errors for the user, so
we remove that in this commit.